Merge

+++ src/hotspot/share/gc/shared/stringdedup/stringDedupTable.hpp
@@ -131,7 +131,7 @@ private:
   // zero hash seed means we will use the Java compatible hash
   // function (which doesn't use a seed), and a non-zero hash
   // seed means we use the murmur3 hash function.
-  jint                            _hash_seed;
+  uint64_t                        _hash_seed;
 
   // Constants governing table resize/rehash/cache.
   static const size_t             _min_size;
@@ -153,7 +153,7 @@ private:
   static StringDedupTable*        _resized_table;
   static StringDedupTable*        _rehashed_table;
 
-  StringDedupTable(size_t size, jint hash_seed = 0);
+  StringDedupTable(size_t size, uint64_t hash_seed = 0);
   ~StringDedupTable();
 
   // Returns the hash bucket at the given index.
